Week 03 Homework
Used floats to create two column layout, Open Sans fonts and styled
them according to the description set in the [design file]. There is
one stylesheet for my css work for both the landing page done last
week, and I have linked that to the blog page I’ve done for this week.
Each link and button is interactive with either hover effects or
javascript. Click ‘About’ on the Blog page to go to the Landing page,
Click ‘Get It Now’ on the landing page to go to return to the ‘Blog’
page. Blog link is anchored to the Blog page.